+++ components/bridge/impl/src/main/java/com/flipperdevices/bridge/impl/manager/service/FlipperInformationApiImpl.kt
@@ -10,11 +10,13 @@ import com.flipperdevices.bridge.impl.di.BridgeImplComponent
 import com.flipperdevices.bridge.impl.manager.UnsafeBleManager
 import com.flipperdevices.core.di.ComponentHolder
 import com.flipperdevices.core.log.LogTagProvider
+import com.flipperdevices.core.log.info
 import com.flipperdevices.core.preference.pb.PairSettings
 import com.flipperdevices.shake2report.api.Shake2ReportApi
 import java.util.UUID
 import javax.inject.Inject
 import javax.inject.Provider
+import kotlin.experimental.and
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.MutableStateFlow
@@ -24,6 +26,7 @@ import kotlinx.coroutines.flow.onEach
 import kotlinx.coroutines.flow.update
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.plus
+import no.nordicsemi.android.ble.data.Data
 
 private const val MAX_BATTERY_LEVEL = 100
 
@@ -110,15 +113,57 @@ class FlipperInformationApiImpl(
         bleManager.readCharacteristicUnsafe(
             infoCharacteristics[Constants.BatteryService.BATTERY_LEVEL]
         ).with { _, data ->
-            val content = data.value ?: return@with
-            val batteryLevel = content.firstOrNull() ?: return@with
-            val batteryLevelAsFloat = batteryLevel.toFloat() / MAX_BATTERY_LEVEL
-            if (batteryLevelAsFloat in 0.0f..1.0f) {
-                informationState.update {
-                    it.copy(batteryLevel = batteryLevelAsFloat)
-                }
-            }
+            onBatteryLevelReceived(data)
         }.enqueue()
+        bleManager.setNotificationCallbackUnsafe(
+            infoCharacteristics[Constants.BatteryService.BATTERY_LEVEL]
+        ).with { _, data ->
+            onBatteryLevelReceived(data)
+        }
+        bleManager.enableNotificationsUnsafe(
+            infoCharacteristics[Constants.BatteryService.BATTERY_LEVEL]
+        ).enqueue()
+
+        bleManager.readCharacteristicUnsafe(
+            infoCharacteristics[Constants.BatteryService.BATTERY_POWER_STATE]
+        ).with { _, data ->
+            onBatteryPowerStateReceived(data)
+        }.enqueue()
+        bleManager.setNotificationCallbackUnsafe(
+            infoCharacteristics[Constants.BatteryService.BATTERY_POWER_STATE]
+        ).with { _, data ->
+            onBatteryPowerStateReceived(data)
+        }
+        bleManager.enableNotificationsUnsafe(
+            infoCharacteristics[Constants.BatteryService.BATTERY_POWER_STATE]
+        ).enqueue()
+    }
+
+    private fun onBatteryLevelReceived(data: Data) {
+        val content = data.value ?: return
+        val batteryLevel = content.firstOrNull() ?: return
+        info { "Battery level is $batteryLevel" }
+        val batteryLevelAsFloat = batteryLevel.toFloat() / MAX_BATTERY_LEVEL
+        if (batteryLevelAsFloat in 0.0f..1.0f) {
+            informationState.update {
+                it.copy(batteryLevel = batteryLevelAsFloat)
+            }
+        }
+    }
+
+    private fun onBatteryPowerStateReceived(data: Data) {
+        val content = data.value ?: return
+        val batteryPowerState = content.firstOrNull() ?: return
+        info { "Battery power state is $batteryPowerState" }
+        if (batteryPowerState and Constants.BatteryService.BATTERY_POWER_STATE_MASK ==
+            Constants.BatteryService.BATTERY_POWER_STATE_MASK
+        ) {
+            informationState.update {
+                it.copy(isCharging = true)
+            }
+        } else informationState.update {
+            it.copy(isCharging = false)
+        }
     }
 
     private fun onDeviceNameReceived(deviceName: String) {