+++ components/info/impl/src/main/java/com/flipperdevices/info/impl/compose/info/ComposableFirmwareVersion.kt
@@ -1,8 +1,10 @@
 package com.flipperdevices.info.impl.compose.info
 
 import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
 import com.flipperdevices.info.impl.R
+import com.flipperdevices.info.shared.ComposableDeviceInfoRow
+import com.flipperdevices.info.shared.ComposableDeviceInfoRowText
+import com.flipperdevices.info.shared.ComposableDeviceInfoRowWithText
 import com.flipperdevices.info.shared.getColorByChannel
 import com.flipperdevices.info.shared.getTextByVersion
 import com.flipperdevices.updater.model.FirmwareVersion
@@ -24,7 +26,11 @@ fun ComposableFirmwareVersion(
         R.string.info_device_info_version,
         firmwareVersionInProgress
     ) {
-        ComposableFirmwareVersionValue(it, firmwareVersion)
+        ComposableDeviceInfoRowText(
+            modifier = it,
+            text = getTextByVersion(firmwareVersion),
+            colorId = getColorByChannel(firmwareVersion.channel)
+        )
     }
 }
 
@@ -46,17 +52,5 @@ fun ComposableFirmwareBuildDate(
         R.string.info_device_info_build_date,
         firmwareVersionInProgress,
         firmwareVersion.buildDate
-    )
-}
-
-@Composable
-fun ComposableFirmwareVersionValue(
-    modifier: Modifier = Modifier,
-    version: FirmwareVersion
-) {
-    ComposableDeviceInfoRowText(
-        modifier,
-        text = getTextByVersion(version),
-        colorId = getColorByChannel(version.channel)
     )
 }