+++ components/settings/impl/src/main/java/com/flipperdevices/settings/impl/composable/elements/SwitchableElement.kt
@@ -4,7 +4,6 @@ import androidx.annotation.StringRes
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.interaction.MutableInteractionSource
 import androidx.compose.foundation.layout.Row
-import androidx.compose.material.Switch
 import androidx.compose.material.ripple.rememberRipple
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.remember
@@ -12,7 +11,6 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 
 @Composable
-@Suppress("LongParameterList")
 fun SwitchableElement(
     @StringRes titleId: Int? = null,
     @StringRes descriptionId: Int? = null,