+++ components/updater/card/src/main/java/com/flipperdevices/updater/card/api/UpdaterCardApiImpl.kt
@@ -1,0 +1,17 @@
+package com.flipperdevices.updater.card.api
+
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import com.flipperdevices.core.di.AppGraph
+import com.flipperdevices.updater.api.UpdaterCardApi
+import com.flipperdevices.updater.card.composable.ComposableUpdaterCardInternal
+import com.squareup.anvil.annotations.ContributesBinding
+import javax.inject.Inject
+
+@ContributesBinding(AppGraph::class)
+class UpdaterCardApiImpl @Inject constructor() : UpdaterCardApi {
+    @Composable
+    override fun ComposableUpdaterCard(modifier: Modifier) {
+        ComposableUpdaterCardInternal(modifier)
+    }
+}